move types to separate file (#1300)
* move types to separate file #1266
* fix relative files #1266
