+++ packages/docs/src/components/DocsTable/DocsTable.ts
@@ -1,4 +1,1 @@
-export type TableData = string[][];
-export type TableColumn = string | { title: string, type?: 'strong' | 'markdown' | 'code' | 'pre' | 'plain' | 'translationString' }
-
 export { default as DocsTable } from './DocsTable.vue'
