+++ packages/docs/src/components/DocsTable/DocsTableTypes.ts
@@ -1,0 +1,2 @@
+export type TableData = string[][];
+export type TableColumn = string | { title: string, type?: 'strong' | 'markdown' | 'code' | 'pre' | 'plain' | 'translationString' }
