+++ packages/docs/src/helpers/DocsHelper.ts
@@ -10,7 +10,7 @@ import {
 } from '../types/configTypes'
 import { DefineComponent } from 'vue'
 import { VueConstructor } from 'vue-class-component'
-import { TableData, TableColumn } from './../components/DocsTable/DocsTable'
+import { TableData, TableColumn } from './../components/DocsTable/DocsTableTypes'
 
 export class PageGenerationHelper {
   path: string
