+++ packages/docs/src/page-configs/services/global-config/page-config.ts
@@ -6,7 +6,7 @@ import {
   reactiveSetExampleCode,
   useInRuntimeCode,
 } from './code-examples'
-import { TableColumn, TableData } from '../../../components/DocsTable/DocsTable'
+import { TableColumn, TableData } from '../../../components/DocsTable/DocsTableTypes'
 
 const columns: TableColumn[] = [
   'params',
